# feat(tableutil): add internal table utility for streaming tables