This creates a new utility for constructing a table stream. Its primary
focus is to allow sources that create tables to easily create buffered
table streams. This is demonstrated by updating the internal point
generator so that it will now natively construct arrow arrays using the
arrow table buffer and stream the chunks downstream using the stream
method.

These utilities are being kept as internal packages while we experiment
with possible methods of constructing tables quickly and efficiently.

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written