@jsternberg 2019-11-11
I forgot to include the benchmark change for this. I wrote a benchmark for `gen.Input` to show the change.

```
benchmark                     old ns/op       new ns/op     delta
BenchmarkInput/1000-12        57067289        12255450      -78.52%
BenchmarkInput/100000-12      6977867133      19307882      -99.72%
BenchmarkInput/1000000-12     64813674783     21678660      -99.97%

benchmark                     old allocs     new allocs     delta
BenchmarkInput/1000-12        420137         8436           -97.99%
BenchmarkInput/100000-12      40033514       10735          -99.97%
BenchmarkInput/1000000-12     400040503      10735          -100.00%

benchmark                     old bytes       new bytes     delta
BenchmarkInput/1000-12        41218925        7172487       -82.60%
BenchmarkInput/100000-12      7173513832      16520755      -99.77%
BenchmarkInput/1000000-12     68339805312     16558400      -99.98%
```