@jsternberg 2019-11-12
The code for the benchmark is part of the PR. It is in `internal/gen/input_test.go`.